Second Commit

+++ entity/wallet.go
@@ -2,5 +2,5 @@ package entity
 
 type Wallet struct{
 	Mobile string `json:"mobile"`
-	Balance int64 `json:"balance"`
+	Balance float64 `json:"balance"`
 }
